requirements.txt
black
flake8
isort
mkdocs
pip
python-dotenv
tqdm
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
nltk
matplotlib
torch
notebook
PyYAML
python-box==6.0.2
fastapi==0.78.0
uvicorn==0.18.3
ensure==1.0.2
-e .
